requirements.txt
annotated-types==0.6.0
anyio==4.3.0
bcrypt==4.1.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.4
decorator==5.1.1
Deprecated==1.2.14
fabric==3.2.2
fastapi==0.109.2
h11==0.14.0
idna==3.6
invoke==2.2.0
paramiko==3.4.0
pycparser==2.21
pydantic==2.6.1
pydantic_core==2.16.2
PyJWT==2.8.0
PyNaCl==1.5.0
python-dotenv==1.0.1
python-vagrant==1.0.0
requests==2.31.0
sniffio==1.3.0
starlette==0.36.3
typing_extensions==4.9.0
urllib3==2.2.1
uvicorn==0.27.1
wrapt==1.16.0
